refactor(cli): drain mcp/_legacy.rs and split oversized files (issue #70)
- Move residual tests from _legacy.rs to their domain modules:
  startup_banner → banner.rs, replay_export → handlers/record.rs,
  click_tool_sends_json_rpc_request → handlers/interact.rs,
  auto_detected_socket_is_pinned → server/tests.rs.
- Delete _legacy.rs.
- Split tools/mod.rs (243→65): extract tests to tools/tests.rs +
  tools/schema_tests.rs.
- Split tools/schemas.rs (283): per-domain submodules under
  tools/schemas/ (core, interact, eval, observe, storage, assert,
  record) with flat re-exports in schemas/mod.rs.
- Split server/methods.rs (266): server/calls.rs (target_call,
  call_logs_tool, call_network_tool, assert_count, assert_url),
  server/asserts.rs (assert_text, assert_value, assert_bool),
  server/file_ops.rs (call_drop_tool, call_replay_tool).

Final mcp/ tree: every file ≤ 150 lines. Allow count unchanged at 10
(Task 7 already killed the last too_many_lines in mcp).
