+++ crates/tauri-pilot-cli/src/mcp/handlers/interact.rs
@@ -1,5 +1,5 @@
 //! Handlers for the `interact` tool group: click, fill, type, press, select,
-//! check, scroll, drag, drop.
+//! check, scroll, drag, drop. Tests in `interact_tests.rs`.
 
 use rmcp::{
     ErrorData as McpError,
@@ -83,5 +83,56 @@ pub(super) async fn dispatch(
         }
         "drop" => server.call_drop_tool(args.clone(), window).await,
         _ => unreachable!("handlers/mod.rs guarantees prefix match: {name}"),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use serde_json::{Map, json};
+    use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
+    use tokio::net::UnixListener;
+
+    use super::super::super::server::PilotMcpServer;
+    use super::super::call_tool_by_name;
+    use crate::protocol::{Request, Response};
+
+    #[tokio::test]
+    async fn click_tool_sends_json_rpc_request() {
+        let socket = std::env::temp_dir().join(format!(
+            "tauri-pilot-mcp-click-test-{}.sock",
+            std::process::id()
+        ));
+        let _ = std::fs::remove_file(&socket);
+        let listener = UnixListener::bind(&socket).expect("bind mock socket");
+        let server = tokio::spawn(async move {
+            let (stream, _) = listener.accept().await.expect("accept");
+            let (reader, mut writer) = stream.into_split();
+            let mut reader = BufReader::new(reader);
+            let mut line = String::new();
+            reader.read_line(&mut line).await.expect("read request");
+            let request: Request = serde_json::from_str(line.trim()).expect("parse request");
+            assert_eq!(request.method, "click");
+            assert_eq!(request.params, Some(json!({"ref": "e3"})));
+            let mut response =
+                serde_json::to_vec(&Response::success(request.id, json!({"ok": true})))
+                    .expect("serialize response");
+            response.push(b'\n');
+            writer.write_all(&response).await.expect("write response");
+        });
+
+        let pilot = PilotMcpServer::new(Some(socket.clone()), None);
+        let mut args = Map::new();
+        args.insert("target".to_owned(), json!("@e3"));
+        let result = call_tool_by_name(&pilot, "click", args)
+            .await
+            .expect("tool call succeeds");
+        assert_eq!(result.is_error, Some(false));
+        assert_eq!(
+            result.structured_content,
+            Some(json!({"result": {"ok": true}}))
+        );
+
+        server.await.expect("mock server task");
+        let _ = std::fs::remove_file(&socket);
     }
 }
