+++ crates/tauri-pilot-cli/src/mcp/tools/mod.rs
@@ -60,184 +60,6 @@ fn spec_to_tool(spec: ToolSpec) -> Tool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Every tool registered in `tools()` has a unique name.
-    #[test]
-    fn test_tools_registry_has_unique_names() {
-        let tools = tools();
-        let mut names: Vec<&str> = tools.iter().map(|t| t.name.as_ref()).collect();
-        names.sort_unstable();
-        let total = names.len();
-        names.dedup();
-        assert_eq!(
-            total,
-            names.len(),
-            "duplicate tool names: total={total}, unique={}",
-            names.len()
-        );
-    }
-
-    /// All known tool names (canonical list at PR1 baseline). Must stay stable
-    /// across the upcoming Task 4–8 module split — additions/removals require
-    /// updating BOTH this list and the `tools()` registry in lock-step.
-    #[test]
-    fn test_tools_registry_contains_baseline_tools() {
-        let tools = tools();
-        let names: std::collections::HashSet<&str> =
-            tools.iter().map(|t| t.name.as_ref()).collect();
-
-        let expected = [
-            // core
-            "ping",
-            "windows",
-            "state",
-            "snapshot",
-            "diff",
-            "screenshot",
-            "navigate",
-            "url",
-            "title",
-            "wait",
-            // interact
-            "click",
-            "fill",
-            "type",
-            "press",
-            "select",
-            "check",
-            "scroll",
-            "drag",
-            "drop",
-            // inspect
-            "text",
-            "html",
-            "value",
-            "attrs",
-            // eval
-            "eval",
-            "ipc",
-            // observe
-            "watch",
-            "logs",
-            "network",
-            // storage
-            "storage_get",
-            "storage_set",
-            "storage_list",
-            "storage_clear",
-            "forms",
-            // assert
-            "assert_text",
-            "assert_contains",
-            "assert_visible",
-            "assert_hidden",
-            "assert_value",
-            "assert_count",
-            "assert_checked",
-            "assert_url",
-            // record
-            "record_start",
-            "record_stop",
-            "record_status",
-            "replay",
-        ];
-
-        for name in expected {
-            assert!(
-                names.contains(name),
-                "expected tool '{name}' missing from registry"
-            );
-        }
-    }
-
-    /// Tool count matches baseline. If you add a tool, bump this number AND
-    /// add it to `test_tools_registry_contains_baseline_tools`.
-    #[test]
-    fn test_tools_registry_count_baseline() {
-        let count = tools().len();
-        // Snapshot of pre-split registry. If you intentionally add a tool, update.
-        assert_eq!(count, 45, "tool count drifted from baseline");
-    }
-
-    #[test]
-    fn tool_list_is_sorted_alphabetically() {
-        let tools = tools();
-        let names: Vec<&str> = tools.iter().map(|tool| tool.name.as_ref()).collect();
-        let expected = vec![
-            "assert_checked",
-            "assert_contains",
-            "assert_count",
-            "assert_hidden",
-            "assert_text",
-            "assert_url",
-            "assert_value",
-            "assert_visible",
-            "attrs",
-            "check",
-            "click",
-            "diff",
-            "drag",
-            "drop",
-            "eval",
-            "fill",
-            "forms",
-            "html",
-            "ipc",
-            "logs",
-            "navigate",
-            "network",
-            "ping",
-            "press",
-            "record_start",
-            "record_status",
-            "record_stop",
-            "replay",
-            "screenshot",
-            "scroll",
-            "select",
-            "snapshot",
-            "state",
-            "storage_clear",
-            "storage_get",
-            "storage_list",
-            "storage_set",
-            "text",
-            "title",
-            "type",
-            "url",
-            "value",
-            "wait",
-            "watch",
-            "windows",
-        ];
-        assert_eq!(names, expected);
-    }
-
-    #[test]
-    fn schemas_include_window_override() {
-        use super::super::schemas::target_schema;
-        use serde_json::Value;
-
-        let schema = target_schema();
-        let properties = schema
-            .get("properties")
-            .and_then(Value::as_object)
-            .expect("schema has properties");
-        assert!(properties.contains_key("target"));
-        assert!(properties.contains_key("window"));
-    }
-
-    #[test]
-    fn windows_schema_omits_window_override() {
-        use serde_json::Value;
-
-        let schema = core::global_empty_schema();
-        let properties = schema
-            .get("properties")
-            .and_then(Value::as_object)
-            .expect("schema has properties");
-        assert!(!properties.contains_key("window"));
-    }
-}
+mod schema_tests;
+#[cfg(test)]
+mod tests;
